hey it's compiled!

+++ code/headers_GPU/jitteredGrid_GPU.h
@@ -8,4 +8,5 @@
 // Coord getClosestPoint(const Grid2D & grid, const LavaCake::vec3f & point, const uint32_t gridLayer);
 
 void generateGrid_GPU(uint16_t subdivision, int seed, int gridLayer,std::string filename, int &point_index);
-__global__ void generateCells_GPU(uint16_t init_subdiv, bool isTextureUsed, int* density_images);
+
+__global__ void generateCells_GPU(uint16_t init_subdiv, bool isTextureUsed, float* density_images);
